refactor: move OpenSearchConnector to common package
In Tasklist, the ElasticsearchConnector and OpenSearchConnector were
residing in different packages:
- generic `io.camunda.tasklist.connect` for Elasticsearch
- specific `io.camunda.tasklist.os` for OpenSearch

They are now in the same generic `io.camunda.tasklist.connect` package.

This refactoring allows the methods in `OpenSearchConnector` annotated
`@VisibleForTesting` to no longer be public, since the class and the
tests are in the same package.
