+++ qa/archunit-tests/src/test/java/io/camunda/VisibleForTestingArchTest.java
@@ -51,9 +51,6 @@ public class VisibleForTestingArchTest {
           "io.camunda.db.rdbms.write.service.HistoryCleanupService.resolveBatchOperationTTL(io.camunda.search.entities.BatchOperationType)",
           "io.camunda.exporter.handlers.AuditLogHandler.getTransformer()",
           "io.camunda.exporter.rdbms.handlers.AuditLogExportHandler.getTransformer()",
-          "io.camunda.tasklist.os.OpenSearchConnector.isHealthy(org.opensearch.client.opensearch.OpenSearchClient)",
-          "io.camunda.tasklist.os.OpenSearchConnector.setTasklistObjectMapper(com.fasterxml.jackson.databind.ObjectMapper)",
-          "io.camunda.tasklist.os.OpenSearchConnector.setTasklistProperties(io.camunda.tasklist.property.TasklistProperties)",
           "io.camunda.zeebe.backup.client.api.BackupRequestHandler.takeBackup(long, io.camunda.zeebe.protocol.record.value.management.CheckpointType)",
           "io.camunda.zeebe.broker.exporter.repo.ExporterRepository.validateAndAddExporterDescriptor(java.lang.String, java.lang.Class, java.util.Map)",
           "io.camunda.zeebe.broker.system.partitions.impl.AsyncSnapshotDirector.getCommitPosition()",
